perf(test): reduce test startup overhead
