+++ src/config/config.broadcast.test.ts
@@ -1,9 +1,8 @@
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
+import { validateConfigObject } from "./config.js";
 
 describe("broadcast", () => {
-  it("accepts a broadcast peer map with strategy", async () => {
-    vi.resetModules();
-    const { validateConfigObject } = await import("./config.js");
+  it("accepts a broadcast peer map with strategy", () => {
     const res = validateConfigObject({
       agents: {
         list: [{ id: "alfred" }, { id: "baerbel" }],
@@ -16,18 +15,14 @@ describe("broadcast", () => {
     expect(res.ok).toBe(true);
   });
 
-  it("rejects invalid broadcast strategy", async () => {
-    vi.resetModules();
-    const { validateConfigObject } = await import("./config.js");
+  it("rejects invalid broadcast strategy", () => {
     const res = validateConfigObject({
       broadcast: { strategy: "nope" },
     });
     expect(res.ok).toBe(false);
   });
 
-  it("rejects non-array broadcast entries", async () => {
-    vi.resetModules();
-    const { validateConfigObject } = await import("./config.js");
+  it("rejects non-array broadcast entries", () => {
     const res = validateConfigObject({
       broadcast: { "120363403215116621@g.us": 123 },
     });
