+++ src/config/config.msteams.test.ts
@@ -1,9 +1,8 @@
-import { describe, expect, it, vi } from "vitest";
+import { describe, expect, it } from "vitest";
+import { validateConfigObject } from "./config.js";
 
 describe("config msteams", () => {
-  it("accepts replyStyle at global/team/channel levels", async () => {
-    vi.resetModules();
-    const { validateConfigObject } = await import("./config.js");
+  it("accepts replyStyle at global/team/channel levels", () => {
     const res = validateConfigObject({
       channels: {
         msteams: {
@@ -29,9 +28,7 @@ describe("config msteams", () => {
     }
   });
 
-  it("rejects invalid replyStyle", async () => {
-    vi.resetModules();
-    const { validateConfigObject } = await import("./config.js");
+  it("rejects invalid replyStyle", () => {
     const res = validateConfigObject({
       channels: { msteams: { replyStyle: "nope" } },
     });
