+++ test/setup.ts
@@ -162,7 +162,6 @@ beforeEach(() => {
 });
 
 afterEach(() => {
-  setActivePluginRegistry(createDefaultRegistry());
   // Guard against leaked fake timers across test files/workers.
   vi.useRealTimers();
 });
